refactor(state): align inverse calc naming and collapse activeTab mutation (#692)
Split the combined inverse-calc.svelte.ts into one-effect-per-file
inverse-range-calc.svelte.ts + inverse-stp-calc.svelte.ts so the inverse
family matches the multi-*-calc convention (data in inverse-lookups, one
calculation $effect per sibling file). Removed the redundant
`set activeTab(v)` setter, leaving setActiveTab(tab) as the single
mutation entry point and marking the interface field readonly.

No behaviour change; inverse-lookup tests unchanged.
